+++ site/git/repos.nix
@@ -1,5 +1,23 @@
 [
   {
+    name = "dotfiles";
+    url = "https://github.com/maydayv7/dotfiles";
+    description = "My PC Dotfiles & Website";
+    owner = "V7";
+  }
+  {
+    name = "stagit";
+    url = "https://github.com/maydayv7/stagit";
+    description = "Personal Stagit Fork";
+    owner = "V7";
+  }
+  {
+    name = "creek";
+    url = "https://github.com/maydayv7/creek";
+    description = "The AI Image Editor of 2030";
+    owner = "V7";
+  }
+  {
     name = "matiru";
     url = "https://github.com/maydayv7/matiru";
     description = "Blockchain-Based Supply Chain Transparency for Agricultural Produce";